+++ dgisim/src/character/characters.py
@@ -48,8 +48,12 @@ class Characters:
 
     def get_alive_character_in_activity_order(self) -> tuple[Character, ...]:
         for i, character in enumerate(self._characters):
-            if character.get_id() == self._active_character_id and character.alive():
-                return self._characters[i:] + self._characters[:i]
+            if character.get_id() == self._active_character_id:
+                return tuple(
+                    char
+                    for char in (self._characters[i:] + self._characters[:i])
+                    if char.alive()
+                )
         return self._characters
 
     def get_character_ordered_from_id(self, char_id: int) -> tuple[Character, ...]: